src/app/layouts/base-layout/base-layout.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { By } from '@angular/platform-browser';
import { RouterModule } from '@angular/router';
import { BaseLayoutComponent } from './base-layout.component';

describe(BaseLayoutComponent.name, () => {
  let component: BaseLayoutComponent;
  let fixture: ComponentFixture<BaseLayoutComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [
        // Component under test
        BaseLayoutComponent,

        // Angular dependencies
        RouterModule.forRoot([]),
      ],
    }).compileComponents();

    fixture = TestBed.createComponent(BaseLayoutComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it.each(['app-vert-menu'])('%s', (selector) => {
    expect(fixture.debugElement.query(By.css(selector))).toBeTruthy();
  });
});
